Add DB subscription (#114)
Adds a struct that can subscribe to changes in a database query,
starting from some cursor onwards. Features:
- Poll based on an interval or on an input notifier channel
- Serve updates via an output channel
- Stops polling when context is closed
- Non-racy unit tests (hopefully)

Also adds a flexible DB query against the gateway_envelopes table that
can query based on a range of params. Given the size of the query, I
should add a unit test for it eventually.