Add RegisterNode function (#143)
## tl;dr

- Adds `NodeRegistryAdmin` struct to handle administrative actions on the Nodes smart contract
- Makes it possible to deploy fresh smart contracts in tests, so we can execute against a clean contract state

## More info

We have one blockchain node shared in the test environment, but we are going to want to be able to write tests that run against the same smart contract without interfering with one another.

To handle that we now have a `deploy-ephemeral` script which will create a new instance of the smart contract at a new address. We can call that from inside tests to get a clean contract to test against.

I still have the first deployed contract being loaded from the `_build` folder, which is being used in some older tests. We may want to phase that out and just always work against clean smart contracts.