Add coroutine labels (#258)
This is a little change that makes using the Go debugger so much easier.
You can identify the goroutine by its name and other additional labels.

We don't always use the tracing tool to start our goroutines, but we do
the major ones. We can always expand this to other locations that would
benefit from easy identification.