Handle registry changes (#252)
There are some drive-by changes to how we log closing events. It was
info, which is okay, but you don't usually need to print 30 lines when
shutting down a service. So I moved all of them to debug. You need to be
actively debugging shutdown issues to need those.

I fixed shutdown. It was pretty broken. The large issue was that we were
waiting on server shutdown in various blocking methods before we
actually called the context cancel method. Now we first cancel the
context and then go into the various shutdown routines. This, amongst
other improvements, means that shutdown now doesn't have to timeout and
is much faster.

The meat of this changeset is the introduction of registry listeners.
The challenge there is that we don't hold a lock on the registry between
calls so I had to refactor a few things to be thread safe. I always
chose the pattern of establishing a channel/registry/listener first,
then getting the newest state, then acting on that state.

Another annoying part is that we can only rebuild the sync client by
cancelling the context. Otherwise it will just sit in Recv() and never
wake up. This means that it can't be a simple channel and it was a
slightly bigger surgery.

I have tested this manually and it behaves as expected. It needs a
follow up combo of unit tests. We also probably need to introduce larger
node2node tests. The blockchain is very slow, changes take solid 20+
seconds, so its not practical for integration tests. Hopefully we can
just use the mock registry.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
- Added subscription management improvements in the `syncWorker` for
better handling of node subscriptions.

- **Bug Fixes**
- Enhanced error handling during the shutdown process of the
`ReplicationServer`.

- **Refactor**
- Simplified the `get-all-nodes` command by removing unnecessary
parameters.
- Updated logging levels across various components to improve verbosity
and clarity.

- **Tests**
- Added a skip message to the `TestCreateServer` function indicating
that the test is currently broken.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->