Terminate self on prolonged chain disconnect (#271)
If the chain is unreachable during startup, the node will refuse to come
up since it can't determine its nodeid.
If the chain goes away (for example due to a partial network failure)
the node can drift too far away from the ordered MLS messages. We can
prevent that scenario from happening by shutting ourselves down.

There are generally 3 ways to shut down a node:
- cancel the main context
- signal self
- panic

I deliberately picked `panic` here as it is the most visible of the
three. It will print a stack trace which usually indicates that
something is not ok. Its much more "visible" than an error message and a
clean shut down.

An orchestrator will restart the node but if the disconnect persists, it
will not come up.

The default self termination timeout is 5 min during which we will
attempt to re-connect every 100ms. A network that does not heal within
5min sounds like a larger outage.