Sensible DB wait timeout (#407)
In #405 I fixed the DB timeout. But not all environments override the
variable.

In some cases this means that the timeout is now `0` and fails
immediately. 0 does not mean "context with infinite timeout" it means
"context that is already timed out"

Set it to 30s which sounds reasonable.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Configuration**
- Added a default 30-second wait time for database connection
initialization when no specific duration is specified.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->