SQL Nonce Manager for HA Payer (#578)
Create a SQL backed nonce manager so we can have HA payer using the same
wallet.

This uses SQL. We know we can't trust PostgreSQL sequences to generate
gapless sequences, which is a requirement for nonces. So, this generates
a table of nonces that the various threads can consume.
- one thread generates nonces 10k at a time. I suspect we might need to
adjust this value over time.
- many threads lock, attempt to publish, commit/consume

SQL guarantees that if the payer fails, or Cancel() gets called, the
nonce is released back into the pool. A future transaction can grab that
nonce again.

There is a lot of SQL magic here using `FOR UPDATE SKIP LOCKED`.

The nonce release/consume is kinda convoluted so its extracted into its
own `withNonce` function.

This manager has a at-least-once semantic. It will never skip a nonce
and ideal cases will generate each nonce exactly once. But during
failure, it might return a nonce more than once. This might lead to
"underpriced" and "nonce too low" errors. We can internally handle those
without exposing them to the client. If "nonce too low" is detected, we
just attempt to use the next one until we get something usable.

The payer now needs a DB. I am just using the same one... In our
environments, we can point the payer at the same DB instance and
migration will take care of things. In other environments, a separate DB
can be provided. I think that it is a lot of work to introduce a
completely separate DB with separate migrations, so I opted for this
quick and dirty solution.

There are a lot of unit tests, but I have to deliver this in stages.
- this PR
- get this into staging
- stress test this
- probably fix a few things here and there

I have raised the bar on test flakiness. I am trying to get the tests to
pass for 10+ min of running continuously. We are not there yet, but at
least the blockchain related tests don't fail. There are lots of changes
to tests in this PR that came from that effort. Mostly retries.