IPayerReport interface (#581)
Context around Merkle Trees in https://github.com/xmtp/xmtpd/issues/563
Closes https://github.com/xmtp/xmtpd/issues/470?issue=xmtp%7Cxmtpd%7C499

The following is the expected sequence of events:
- A node generates and submits a `PayerReport`.
- All the other nodes call `attestPayerReport` if they can confirm its
validity.
- Nodes in the network listen for attestations and tracks them. Upon
reaching majority, the `confirmPayerReport` can be called.
- When a `PayerReport` is confirmed, any node can submit batches to
`settleUsageBatch`. The `PayerReport` contract is expected to track
batches submissions to prevent duplicates.
- `settleUsageBatch` verifies the provided payers and amounts leafs are
included in the PayerReport Merkle tree. If verified, the
Payer.settleUsage is called with the list of payers and amounts to
settle the usage.

Nodes attest to PayerReports, and once majority has been reached
`confirmPayerReport` can be called for settlement.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
- Enhanced the usage settlement process to handle batches more
efficiently with improved verification.
- Introduced a new reporting interface enabling users to submit, attest,
and confirm usage data reports.
- Added events for tracking report submissions, attestations, and
confirmations.
- Updated terminology for contracts related to distribution to better
reflect current functionality.
- **Bug Fixes**
- Clarified documentation for events to accurately represent their
context and usage.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->