Do not register same pubkey for multiple nodes (#692)
In general, we do not want to re-register the same pubkey with a new
NodeId.

If this is truly desired, one can register override with `--force`

This is super useful for test environments when we want a node to
self-register with the registry on startup.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced node registration to automatically prevent duplicate
registrations based on public key matching.
- Introduced a new "Force" option in the registration command, allowing
users to bypass the duplicate check if needed.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->