Fix originator in indexer storer (#1009)
### Fix originator parameter passing in indexer storer
`buildOriginatorEnvelope` functions for GroupMessageStorer and
IdentityUpdateStorer
The `buildOriginatorEnvelope` functions in both
[group_message_storer.go](https://github.com/xmtp/xmtpd/pull/1009/files#diff-640e55965cfecb771c26ece72f66b6e0298881dc9ec9ebbb0dac81aaab3c7274)
and
[identity_update_storer.go](https://github.com/xmtp/xmtpd/pull/1009/files#diff-719fcd3f6b2a547b6c6f299d70464725ddf9c1161c104099b3c3cab140cb2354)
now accept an explicit `originatorId` parameter instead of hardcoding
the originator ID values internally. The corresponding test files
[group_message_storer_test.go](https://github.com/xmtp/xmtpd/pull/1009/files#diff-8513af96262309782bbdafe8c660599bb35200e05e8417dd438b9dc126b4bba0)
and
[identity_update_storer_test.go](https://github.com/xmtp/xmtpd/pull/1009/files#diff-cb367b4ef560a994a2e5938fa5bb995077361165eff69159a3f85fa7a5cae10d)
add assertions to verify that the originator node ID in envelopes
matches the expected constants `GROUP_MESSAGE_ORIGINATOR_ID` and
`IDENTITY_UPDATE_ORIGINATOR_ID` respectively.

#### 📍Where to Start
Start with the `StoreLog` method in
[group_message_storer.go](https://github.com/xmtp/xmtpd/pull/1009/files#diff-640e55965cfecb771c26ece72f66b6e0298881dc9ec9ebbb0dac81aaab3c7274)
to see how the modified `buildOriginatorEnvelope` function is called
with the explicit originator ID parameter.

----

_[Macroscope](https://app.macroscope.com) summarized 028fa37._